modularized login
